consolidate renderer-lib into shared/renderer
Merges @meru/renderer-lib into @meru/shared under the `renderer/`
subpath since both packages existed to hold cross-package helpers with
an artificial split. Switches shared's tsconfig to the dom base with
jsx enabled so it can type-check the moved React/DOM-using modules.
