--- packages/renderer-lib/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "@meru/renderer-lib",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "types": "tsc"
-  }
-}
