--- packages/renderer-lib/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "@total-typescript/tsconfig/bundler/dom",
-  "compilerOptions": {
-    "jsx": "react-jsx",
-    "baseUrl": ".",
-    "paths": {
-      "@/*": ["./*"]
-    }
-  }
-}
