+++ packages/renderer/components/app-titlebar.tsx
@@ -1,4 +1,4 @@
-import { ipc } from "@meru/renderer-lib/ipc";
+import { ipc } from "@meru/shared/renderer/ipc";
 import { accountColorsMap } from "@meru/shared/accounts";
 import { APP_TITLEBAR_HEIGHT, WEBSITE_URL } from "@meru/shared/constants";
 import { type DownloadItem, googleAppsPinnedApps } from "@meru/shared/types";
@@ -27,7 +27,7 @@ import { type ComponentProps, useEffect, useRef, useState } from "react";
 import { useDebouncedCallback } from "use-debounce";
 import { navigate, useHashLocation } from "wouter/use-hash-location";
 import { useIsLicenseKeyValid } from "@/lib/hooks";
-import { useConfig } from "@meru/renderer-lib/react-query";
+import { useConfig } from "@meru/shared/renderer/react-query";
 import {
   useAccountsStore,
   useAppUpdaterStore,
