+++ packages/renderer/components/config-switch-field.tsx
@@ -2,7 +2,7 @@ import type { Config } from "@meru/shared/types";
 import { Field, FieldContent, FieldDescription, FieldLabel } from "@meru/ui/components/field";
 import { Switch } from "@meru/ui/components/switch";
 import { useIsLicenseKeyValid } from "@/lib/hooks";
-import { useConfig, useConfigMutation } from "@meru/renderer-lib/react-query";
+import { useConfig, useConfigMutation } from "@meru/shared/renderer/react-query";
 import { restartRequiredToast } from "@/lib/toast";
 import { LicenseKeyRequiredFieldBadge } from "./license-key-required-field-badge";
