+++ packages/renderer/ipc.ts
@@ -1,4 +1,4 @@
-import { ipc } from "@meru/renderer-lib/ipc";
+import { ipc } from "@meru/shared/renderer/ipc";
 import { playNotificationSound } from "./lib/notifications";
 
 ipc.renderer.on("taskbar.setOverlayIcon", (_event, unreadCount) => {
