+++ packages/renderer/routes/settings/blocker.tsx
@@ -2,7 +2,7 @@ import { FieldGroup, FieldSeparator } from "@meru/ui/components/field";
 import { ConfigSwitchField } from "@/components/config-switch-field";
 import { LicenseKeyRequiredBanner } from "@/components/license-key-required-banner";
 import { Settings, SettingsContent, SettingsHeader, SettingsTitle } from "@/components/settings";
-import { useConfig } from "@meru/renderer-lib/react-query";
+import { useConfig } from "@meru/shared/renderer/react-query";
 
 export function BlockerSettings() {
   const { config } = useConfig();
