+++ packages/renderer/routes/settings/downloads.tsx
@@ -1,4 +1,4 @@
-import { ipc } from "@meru/renderer-lib/ipc";
+import { ipc } from "@meru/shared/renderer/ipc";
 import { Button } from "@meru/ui/components/button";
 import {
   Field,
@@ -11,7 +11,7 @@ import {
 import { Input } from "@meru/ui/components/input";
 import { ConfigSwitchField } from "@/components/config-switch-field";
 import { SettingsHeader, SettingsTitle } from "@/components/settings";
-import { useConfig } from "@meru/renderer-lib/react-query";
+import { useConfig } from "@meru/shared/renderer/react-query";
 import { restartRequiredToast } from "@/lib/toast";
 
 export function DownloadsSettings() {
