+++ packages/shared/renderer/types/window.d.ts
