+++ packages/shared/tsconfig.json
@@ -1,6 +1,7 @@
 {
-  "extends": "@total-typescript/tsconfig/bundler/no-dom",
+  "extends": "@total-typescript/tsconfig/bundler/dom",
   "compilerOptions": {
+    "jsx": "react-jsx",
     "baseUrl": ".",
     "paths": {
       "@/*": ["./*"]
